// src/app/core/api/api.model.ts
export interface QuerySort {
  order: string;
  direction: string;
}

export interface PageableParams {
  page?: number;
  sort?: QuerySort;
  size?: number;
  sortArray?: QuerySort[];
}

export interface PageSlice {
  content: any[];
  last: boolean;
  totalElements: number;
  totalPages: number;
  first: boolean;
  numberOfElements: number;
  sort: any;
  size: number;
  number: number;
}